Tests MSVC build
