+++ test/module/irohad/ametsuchi/mock_temporary_factory.hpp
@@ -1,0 +1,32 @@
+/**
+ * Copyright Soramitsu Co., Ltd. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef IROHA_MOCK_TEMPORARY_FACTORY_HPP
+#define IROHA_MOCK_TEMPORARY_FACTORY_HPP
+
+#include "ametsuchi/temporary_factory.hpp"
+
+#include <gmock/gmock.h>
+
+namespace iroha {
+  namespace ametsuchi {
+
+    class MockTemporaryFactory : public TemporaryFactory {
+     public:
+      MOCK_METHOD0(
+          createTemporaryWsv,
+          expected::Result<std::unique_ptr<TemporaryWsv>, std::string>(void));
+      MOCK_METHOD1(prepareBlock_, void(std::unique_ptr<TemporaryWsv> &));
+
+      void prepareBlock(std::unique_ptr<TemporaryWsv> wsv) override {
+        // gmock workaround for non-copyable parameters
+        prepareBlock_(wsv);
+      }
+    };
+
+  }  // namespace ametsuchi
+}  // namespace iroha
+
+#endif  // IROHA_MOCK_TEMPORARY_FACTORY_HPP
