+++ test/module/irohad/consensus/yac/peer_orderer_test.cpp
@@ -13,7 +13,8 @@
 #include <boost/range/numeric.hpp>
 #include "consensus/yac/impl/peer_orderer_impl.hpp"
 #include "consensus/yac/storage/yac_proposal_storage.hpp"
-#include "module/irohad/ametsuchi/ametsuchi_mocks.hpp"
+#include "module/irohad/ametsuchi/mock_peer_query.hpp"
+#include "module/irohad/ametsuchi/mock_peer_query_factory.hpp"
 #include "module/irohad/consensus/yac/yac_mocks.hpp"
 
 using namespace boost::adaptors;
