+++ test/module/irohad/network/block_loader_test.cpp
@@ -14,7 +14,10 @@
 #include "cryptography/hash.hpp"
 #include "datetime/time.hpp"
 #include "framework/test_subscriber.hpp"
-#include "module/irohad/ametsuchi/ametsuchi_mocks.hpp"
+#include "module/irohad/ametsuchi/mock_block_query.hpp"
+#include "module/irohad/ametsuchi/mock_block_query_factory.hpp"
+#include "module/irohad/ametsuchi/mock_peer_query.hpp"
+#include "module/irohad/ametsuchi/mock_peer_query_factory.hpp"
 #include "module/shared_model/builders/common_objects/peer_builder.hpp"
 #include "module/shared_model/builders/protobuf/common_objects/proto_peer_builder.hpp"
 #include "module/shared_model/builders/protobuf/test_block_builder.hpp"
