+++ test/module/irohad/ordering/on_demand_ordering_gate_test.cpp
@@ -9,7 +9,8 @@
 #include <boost/range/adaptor/indirected.hpp>
 #include "framework/test_subscriber.hpp"
 #include "interfaces/iroha_internal/transaction_batch_impl.hpp"
-#include "module/irohad/ametsuchi/ametsuchi_mocks.hpp"
+#include "module/irohad/ametsuchi/mock_tx_presence_cache.hpp"
+#include "module/irohad/ordering/mock_on_demand_os_notification.hpp"
 #include "module/irohad/ordering/ordering_mocks.hpp"
 #include "module/shared_model/interface_mocks.hpp"
 #include "ordering/impl/on_demand_common.hpp"
