+++ test/module/irohad/torii/torii_mocks.hpp
@@ -19,44 +19,10 @@
 namespace iroha {
   namespace torii {
 
-    class MockQueryProcessor : public QueryProcessor {
-     public:
-      MOCK_METHOD1(queryHandle,
-                   std::unique_ptr<shared_model::interface::QueryResponse>(
-                       const shared_model::interface::Query &));
-      MOCK_METHOD1(
-          blocksQueryHandle,
-          rxcpp::observable<
-              std::shared_ptr<shared_model::interface::BlockQueryResponse>>(
-              const shared_model::interface::BlocksQuery &));
-    };
-
     class MockStatusBus : public StatusBus {
      public:
       MOCK_METHOD1(publish, void(StatusBus::Objects));
       MOCK_METHOD0(statuses, rxcpp::observable<StatusBus::Objects>());
-    };
-
-    class MockCommandServiceTransport
-        : public iroha::protocol::CommandService_v1::Service {
-     public:
-      MOCK_METHOD3(Torii,
-                   grpc::Status(grpc::ServerContext *,
-                                const iroha::protocol::Transaction *,
-                                google::protobuf::Empty *));
-      MOCK_METHOD3(ListTorii,
-                   grpc::Status(grpc::ServerContext *,
-                                const iroha::protocol::TxList *,
-                                google::protobuf::Empty *));
-      MOCK_METHOD3(Status,
-                   grpc::Status(grpc::ServerContext *,
-                                const iroha::protocol::TxStatusRequest *,
-                                iroha::protocol::ToriiResponse *));
-      MOCK_METHOD3(
-          StatusStream,
-          grpc::Status(grpc::ServerContext *,
-                       const iroha::protocol::TxStatusRequest *,
-                       grpc::ServerWriter<iroha::protocol::ToriiResponse> *));
     };
 
     class MockCommandService : public ::torii::CommandService {
