+++ test/module/irohad/validation/chain_validation_test.cpp
@@ -6,7 +6,8 @@
 #include "validation/impl/chain_validator_impl.hpp"
 
 #include <boost/range/adaptor/indirected.hpp>
-#include "module/irohad/ametsuchi/ametsuchi_mocks.hpp"
+#include "module/irohad/ametsuchi/mock_mutable_storage.hpp"
+#include "module/irohad/ametsuchi/mock_peer_query.hpp"
 #include "module/irohad/consensus/yac/yac_mocks.hpp"
 #include "module/shared_model/interface_mocks.hpp"
